Minor refactor
Package location changes and remove redundant code.
